refactor(website): split W-ASAP dashboard page state selector into multiple files
